* SliderComponent works better, can click anywhere on the slider and it's simplified.  CreateSlider changed to work with the new method. * libVLCStreamComponent now has playback/vol/progress controls * ScrollComponent works with the mouse wheel, did that awhile ago but guess it was committed * TouchDragMarkupComponent now recovers correctly after being forced to reload surfaces (like if the screensize changes) * Surface's bmp loader now uses SoftSurface, so it can now can handle non-power of 2 bmps, that was a weird hangover from when only .rttex's were supported * Misc changes and improvements to libVLC_RSP, not fully tested, it's now possible to subscribe to events like pause/init, all stuff to make libVLCStreamComponent work better * Added a GUI slider to the "About" text scrolling example in RTSimpleApp, I don't think an example of how to do a slider control was anywhere else * RTSimpleApp example now has an example GUI slider in it, in the About screen * Drag and drop is now support on Windows, it will create MESSAGE_TYPE_FILE_DROPPED system messages with the x/y and filename.  Check SlideShow for an example of use.  Any down side to just leaving this on for all apps?  I mean, the event will be ignored unless it's actively handled.

+++ shared/Entity/HTTPComponent.cpp
@@ -110,7 +110,7 @@ void HTTPComponent::PrepareConnection(VariantList *pVList)
 
 void HTTPComponent::AddPostData(VariantList *pVList)
 {
-	m_netHTTP.AddPostData(pVList->m_variant[0].GetString(), (uint8*)pVList->m_variant[1].GetString().c_str(), pVList->m_variant[1].GetString().size());
+	m_netHTTP.AddPostData(pVList->m_variant[0].GetString(), (uint8*)pVList->m_variant[1].GetString().c_str(), (int)pVList->m_variant[1].GetString().size());
 }
 
 void HTTPComponent::SetFileOutput(VariantList *pVList)
